GTX-343 separates helper functions into their own files
