+++ skyVegasTests/supportedCurrenciesGameClient.spec.ts
@@ -1,10 +1,8 @@
 import { test, type Page} from "@playwright/test";
-import { validateConsoleMessages, login, launchGame } from "./testHelpers";
+import { launchGame, validateConsoleMessages, login } from "./utilities"
 
 test("supportedCurrencies", async ({page}) => {
     await login(page)
     await launchGame(page, "Big Bass Splash")
     await validateConsoleMessages(page, '"currency":"GBP"')
-
-
 })
