+++ skyVegasTests/utilities/validateConsoleMessagesHelper.ts
@@ -31,23 +31,4 @@ let validateConsoleMessages = async (page: Page, expectedMessage: string) => {
       expect(messageExists).toBe(true);
     };
 
-  let login = async (page:Page) => {
-    const userName = process.env.USERNAME!;
-    const password = process.env.PASSWORD!;
-    const url = process.env.URL!;
-    await page.goto(url!);
-    await page.getByRole("link", { name: "Log In" }).click();
-    await page.getByLabel("Username").fill(userName);
-    await page.getByLabel("PIN").fill(password);
-    await page.getByRole("button", { name: "I Accept" }).click();
-    await page.getByRole("button", { name: "Log in" }).click();
-  };
-
-  let launchGame = async (page: Page, gameName) => {
-    await page.getByText('Search for games...').click();
-    await page.getByPlaceholder('Search for games...').fill(gameName);
-    await new Promise(r => setTimeout(r, 2000));
-    await page.locator('.tile-footer-wrapper').first().click();
-  }
-  
-  export {validateConsoleMessages, login, launchGame};
+    export {validateConsoleMessages}
